// src/enums/amc.enums.ts
export enum AMCInstances {
  INSTANCE_1 = 'Instance 1',
  INSTANCE_2 = 'Instance 2',
  INSTANCE_3 = 'Instance 3',
}

export enum AMCAccessTypes {
  DEFAULT = 'DEFAULT',
  CUSTOM = 'CUSTOM',
}

export enum AMCQueryTypes {
  DATA_QUERY = 'DATA_QUERY',
  AUDIENCE_QUERY = 'AUDIENCE_QUERY',
}

export enum AMCCampaignTypes {
  SP = 'sp',
  SB = 'sb',
  SD = 'sd',
  DSP = 'dsp',
}

export enum AMCCampaignGroupTypes {
  SP = 'sp',
  SB = 'sb',
  SD = 'sd',
  DSP = 'dsp',
  OLV = 'olv',
  STV = 'stv',
}

export enum AMCInstanceRequestStatus {
  INITIATED = 'INITIATED',
  IN_PROGRESS = 'IN PROGRESS',
  AMAZON_PROCESSING = 'AMAZON PROCESSING',
  FULFILLED = 'FULFILLED',
}

export enum AMCScheduleFrequency {
  DAILY = 'Daily',
  WEEKLY = 'Weekly',
}

export enum AMCQueryExecutionType {
  ONCE = 'Once',
  SCHEDULE = 'Schedule',
}

export enum AMCExecutionDays {
  SUNDAY = 'Sunday',
  MONDAY = 'Monday',
  TUESDAY = 'Tuesday',
  WEDNESDAY = 'Wednesday',
  THURSDAY = 'Thursday',
  FRIDAY = 'Friday',
  SATURDAY = 'Saturday',
}

export enum AMCExecutionTime {
  HOUR_0 = 0,
  HOUR_1 = 1,
  HOUR_2 = 2,
  HOUR_3 = 3,
  HOUR_4 = 4,
  HOUR_5 = 5,
  HOUR_6 = 6,
  HOUR_7 = 7,
  HOUR_8 = 8,
  HOUR_9 = 9,
  HOUR_10 = 10,
  HOUR_11 = 11,
  HOUR_12 = 12,
  HOUR_13 = 13,
  HOUR_14 = 14,
  HOUR_15 = 15,
  HOUR_16 = 16,
  HOUR_17 = 17,
  HOUR_18 = 18,
  HOUR_19 = 19,
  HOUR_20 = 20,
  HOUR_21 = 21,
  HOUR_22 = 22,
  HOUR_23 = 23,
}

export enum AMCTypeWindowTypes {
  EXPLICIT = 'EXPLICIT',
}

export enum AMCExecutionCategory {
  AD_HOC = 'AD_HOC',
  SCHEDULED = 'SCHEDULED',
}

export enum InstanceTypeEnum {
  SANDBOX = 'SANDBOX',
  STANDARD = 'STANDARD',
}